+++ packages/plugin-stark-module/package.json
@@ -5,6 +5,7 @@
   "main": "lib/index.js",
   "files": [
     "lib",
+    "templates",
     "test",
     "tsconfig.json"
   ],
