+++ packages/plugin-stark-module/src/entryHelper.ts
