+++ packages/plugin-stark-module/src/setUMDConfig.ts
@@ -1,13 +1,13 @@
 import * as path from 'path';
 import { IPluginAPI } from '@alib/build-scripts';
 import { Options } from './types';
-import { getModules } from './entry.helper';
+import { getModules } from './entryHelper';
 
 interface GetConfig {
   (api: Partial<IPluginAPI>, options?: Options): any;
 }
 
-const getConfig: GetConfig = ({ context, onGetWebpackConfig }, { modules, outputDir, library, flatten }) => {
+const getConfig: GetConfig = ({ context, onGetWebpackConfig }, { modules, outputDir, library, filenameStrategy }) => {
   const { rootDir, userConfig, pkg } = context;
   const { library: userLibrary } = userConfig;
   const { name } = pkg ?? {};
@@ -24,15 +24,17 @@ const getConfig: GetConfig = ({ context, onGetWebpackConfig }, { modules, output
 
     // set umd
     const output = path.resolve(rootDir, outputDir ?? 'dist');
+    console.log('fsfsfsfsa', filenameStrategy ? `${filenameStrategy}.js` : './[name]/index.js');
     config.output
       .path(output)
-      // set output to dist/[name]
-      .filename(flatten ? '[name].js' : './[name]/index.js')
+      // set output to outputDir/[name]
+      .filename(filenameStrategy ? `${filenameStrategy}.js` : './[name]/index.js')
       .library((library || userLibrary || name || 'module') as string)
       .libraryTarget('umd');
 
-    config.plugin('MiniCssExtractPlugin').tap(([args]) => [{ ...args, filename: flatten ? '[name].css' : './[name]/index.css' }]);
+    config.plugin('MiniCssExtractPlugin').tap(([args]) => [{ ...args, filename: filenameStrategy ? `${filenameStrategy}.css` : './[name]/index.css' }]);
 
+    // hack with multi publicpath
     config.devServer.contentBase(path.join(rootDir, 'dist'));
     config.devServer.writeToDisk(true);
   });
