+++ packages/plugin-stark-module/src/types.ts
@@ -1,8 +1,10 @@
 import { IPluginAPI } from '@alib/build-scripts';
 
+type ExternalsObjectElement = Record<'root' | 'amd' | 'commonjs' | 'commonjs2', string>;
+
 export type Depth = 0 | 1 | 2 | 'major' | 'minor' | 'patch';
 
-export interface Runtime {
+export interface Runtime extends Partial<ExternalsObjectElement> {
   root: string;
   id?: string;
   depth?: Depth;
@@ -17,14 +19,13 @@ export type Externals =
   }
 
 export interface Options {
-  // runtime?: Runtime[];
   externals?: Externals;
   modules?: Json<string>[];
   minify?: boolean;
   outputDir?: string;
   sourceMap?: boolean;
   library?: string;
-  flatten?: boolean;
+  filenameStrategy?: string;
 }
 
 export type PartialPlugin = Partial<IPluginAPI>;
