+++ packages/plugin-stark-module/templates/runtime.hbs
